extract parse::dump::Dump;Dump.log_count guarantee use all field
