+++ route_policy_cmp/src/parse/lex.rs
@@ -1,24 +1,15 @@
-use std::{
-    collections::BTreeMap,
-    fs::{create_dir_all, read_dir, File},
-    io::{BufReader, Write},
-    path::Path,
-    thread::available_parallelism,
-};
+use std::collections::BTreeMap;
 
 use anyhow::{bail, Context, Error, Ok, Result};
 use ipnet::IpNet;
-use itertools::izip;
 use lazy_regex::regex_captures;
 use log::{debug, error};
-use rayon::prelude::{
-    IndexedParallelIterator, IntoParallelIterator, IntoParallelRefIterator, ParallelIterator,
-};
-use serde::{Deserialize, Serialize};
+use rayon::prelude::{IntoParallelIterator, ParallelIterator};
 
 use super::{
     aut_num::AutNum,
     aut_sys::{is_as_set, parse_as_name},
+    dump::Dump,
     mp_import::parse_imports,
     peering::{is_peering_set, parse_mp_peering},
     set::{is_route_set_name, AsSet, FilterSet, PeeringSet, RouteSet},
@@ -26,7 +17,6 @@ use super::{
 use crate::{
     lex::{dump, rpsl_object},
     parse::filter::{is_filter_set, parse_filter},
-    serde::from_reader,
 };
 
 pub fn parse_lexed(lexed: dump::Dump) -> Dump {
@@ -39,26 +29,16 @@ pub fn parse_lexed(lexed: dump::Dump) -> Dump {
         filter_sets,
         as_routes,
     } = lexed;
-    let aut_nums = parse_lexed_aut_nums(aut_nums);
-    debug!("Parsed {} Aut Nums.", aut_nums.len());
-    let as_sets = parse_lexed_as_sets(as_sets);
-    debug!("Parsed {} As Sets.", as_sets.len());
-    let route_sets = parse_lexed_route_sets(route_sets);
-    debug!("Parsed {} Route Sets.", route_sets.len());
-    let peering_sets = parse_lexed_peering_sets(peering_sets);
-    debug!("Parsed {} Peering Sets.", peering_sets.len());
-    let filter_sets = parse_lexed_filter_sets(filter_sets);
-    debug!("Parsed {} Filter Sets.", filter_sets.len());
-    let as_routes = parse_lexed_as_routes(as_routes);
-    debug!("Parsed {} AS Routes.", as_routes.len());
-    Dump {
-        aut_nums,
-        as_sets,
-        route_sets,
-        peering_sets,
-        filter_sets,
-        as_routes,
-    }
+    let dump = Dump {
+        aut_nums: parse_lexed_aut_nums(aut_nums),
+        as_sets: parse_lexed_as_sets(as_sets),
+        route_sets: parse_lexed_route_sets(route_sets),
+        peering_sets: parse_lexed_peering_sets(peering_sets),
+        filter_sets: parse_lexed_filter_sets(filter_sets),
+        as_routes: parse_lexed_as_routes(as_routes),
+    };
+    dump.log_count();
+    dump
 }
 
 pub fn parse_lexed_aut_nums(lexed: Vec<rpsl_object::AutNum>) -> BTreeMap<usize, AutNum> {
@@ -212,160 +192,4 @@ pub fn parse_lexed_as_route((name, routes): &(String, Vec<String>)) -> Result<(u
     let num = parse_aut_num_name(name)?;
     let routes: Result<_> = routes.iter().map(|r| Ok(r.parse()?)).collect();
     Ok((num, routes?))
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
-pub struct Dump {
-    pub aut_nums: BTreeMap<usize, AutNum>,
-    pub as_sets: BTreeMap<String, AsSet>,
-    pub route_sets: BTreeMap<String, RouteSet>,
-    pub peering_sets: BTreeMap<String, PeeringSet>,
-    pub filter_sets: BTreeMap<String, FilterSet>,
-    /// The AS numbers with Vec of their routes.
-    /// <https://www.rfc-editor.org/rfc/rfc2622#section-4>.
-    pub as_routes: BTreeMap<usize, Vec<IpNet>>,
-}
-
-pub fn split_n_btreemap<K, V>(mut map: BTreeMap<K, V>, n: usize) -> Vec<BTreeMap<K, V>>
-where
-    K: std::cmp::Ord + Clone,
-{
-    let length = map.len();
-    if length < n {
-        return split_n_btreemap(map, length);
-    }
-
-    let size_per_split = map.len() / n - 1;
-    let mut splits = Vec::with_capacity(n);
-    for _ in 0..(n - 1) {
-        let split_point = map.iter().rev().nth(size_per_split).unwrap().0.clone();
-        let split = map.split_off(&split_point);
-        splits.push(split);
-    }
-    splits.push(map);
-    splits
-}
-
-impl Dump {
-    pub fn split_n(self, n: usize) -> Vec<Self> {
-        let Self {
-            aut_nums,
-            as_sets,
-            route_sets,
-            peering_sets,
-            filter_sets,
-            as_routes,
-        } = self;
-        let aut_num_splits = split_n_btreemap(aut_nums, n);
-        let as_set_splits = split_n_btreemap(as_sets, n);
-        let route_set_splits = split_n_btreemap(route_sets, n);
-        let peering_set_splits = split_n_btreemap(peering_sets, n);
-        let filter_set_splits = split_n_btreemap(filter_sets, n);
-        let as_route_splits = split_n_btreemap(as_routes, n);
-
-        izip!(
-            aut_num_splits,
-            as_set_splits,
-            route_set_splits,
-            peering_set_splits,
-            filter_set_splits,
-            as_route_splits
-        )
-        .map(
-            |(aut_nums, as_sets, route_sets, peering_sets, filter_sets, as_routes)| Self {
-                aut_nums,
-                as_sets,
-                route_sets,
-                peering_sets,
-                filter_sets,
-                as_routes,
-            },
-        )
-        .collect()
-    }
-
-    /// Split self based on the number of CPU logic cores available × 4.
-    pub fn split_n_cpus(self) -> Result<Vec<Self>> {
-        let n: usize = available_parallelism()?.into();
-        Ok(self.split_n(n * 4))
-    }
-
-    /// Quickly write self to `directory` in parallel.
-    pub fn pal_write<P>(self, directory: P) -> Result<()>
-    where
-        P: AsRef<Path>,
-    {
-        let splits = self.split_n_cpus()?;
-        pal_write_dump(&splits, directory)
-    }
-
-    pub fn merge(mut self, other: Self) -> Self {
-        let Self {
-            aut_nums,
-            as_sets,
-            route_sets,
-            peering_sets,
-            filter_sets,
-            as_routes,
-        } = other;
-        self.aut_nums.extend(aut_nums);
-        self.as_sets.extend(as_sets);
-        self.route_sets.extend(route_sets);
-        self.peering_sets.extend(peering_sets);
-        self.filter_sets.extend(filter_sets);
-        self.as_routes.extend(as_routes);
-        self
-    }
-
-    pub fn pal_read<P>(directory: P) -> Result<Self>
-    where
-        P: AsRef<Path>,
-    {
-        let readers = read_dir(directory)?
-            .map(|p| Ok(BufReader::new(File::open(p?.path())?)))
-            .collect::<Result<Vec<_>>>()?;
-        let dumps = readers
-            .into_par_iter()
-            .map(from_reader)
-            .collect::<Result<Vec<_>, _>>()?;
-        Ok(merge_dumps(dumps))
-    }
-
-    pub fn log_count(&self) {
-        debug!(
-            "Parsed {} aut_nums, {} as_sets, {} route_sets, {} peering_sets, {} filter_sets, {} as_routes.",
-            self.aut_nums.len(),
-            self.as_sets.len(),
-            self.route_sets.len(),
-            self.peering_sets.len(),
-            self.filter_sets.len(),
-            self.as_routes.len(),
-        )
-    }
-}
-
-pub fn pal_write_dump<P>(splits: &Vec<Dump>, directory: P) -> Result<()>
-where
-    P: AsRef<Path>,
-{
-    let directory = directory.as_ref().to_owned();
-    create_dir_all(&directory)?;
-    let writes = splits
-        .par_iter()
-        .enumerate()
-        .map(|(index, dump)| {
-            let path = directory.clone().join(format!("{index}.json"));
-            let file = File::create(path)?;
-            let json = serde_json::to_string(dump)?;
-            Ok((file, json))
-        })
-        .collect::<Result<Vec<_>>>()?;
-    for (mut file, json) in writes {
-        file.write_all(json.as_bytes())?;
-    }
-    Ok(())
-}
-
-pub fn merge_dumps(dumps: Vec<Dump>) -> Dump {
-    dumps.into_par_iter().reduce(Dump::default, Dump::merge)
 }
