+++ route_policy_cmp/src/test/parse/lex.rs
@@ -10,8 +10,9 @@ use crate::{
         address_prefix::{AddrPfxRange, RangeOperator::NoOp},
         aut_num::AutNum,
         aut_sys::AsName::*,
+        dump::Dump,
         filter::Filter::*,
-        lex::{parse_aut_num_name, parse_lexed, Dump},
+        lex::{parse_aut_num_name, parse_lexed},
         mp_import::{Casts, Entry, Versions},
         peering::{AsExpr::*, Peering, PeeringAction},
         router_expr::RouterExpr::*,
